Add ALCEExtractor class and integrate into CLI; enhance download method with date parsing and file existence check

+++ src/datapub/entities/al_ce/extractors/diario_extractor.py
@@ -1,14 +1,20 @@
 import json
 import hashlib
 import requests
+import dateparser
+import argparse
+
 from datetime import datetime, date
 
 from datapub.shared.utils.extractor_base import ExtractorBase
 
+
 class ALCEExtractor(ExtractorBase):
     def __init__(self, base_dir="storage/raw/al_ce", extractor_type="diario"):
-        super().__init__(entity="ALCE", base_dir=base_dir, extractor_type=extractor_type)
-        
+        super().__init__(
+            entity="ALCE", base_dir=base_dir, extractor_type=extractor_type
+        )
+
         self.base_api = "https://doalece.al.ce.gov.br/api/publico/ultimas-edicoes"
         self.base_url = "https://doalece.al.ce.gov.br"
 
@@ -19,12 +25,22 @@ class ALCEExtractor(ExtractorBase):
         }
         return f"{self.base_api}?buscarData={json.dumps(date_range)}"
 
+    @staticmethod
+    def add_arguments(parser: argparse.ArgumentParser):
+        parser.add_argument("--start", help="Data inicial no formato YYYY-MM-DD")
+        parser.add_argument("--end", help="Data final no formato YYYY-MM-DD")
+
     def download(self, start=None, end=None):
+        if start is None:
+            start = date(2011, 8, 4)
+        else:
+            start = dateparser.parse(start).date()
+
         if end is None:
             end = date.today()
-        if start is None:
-            start = date(2021, 1, 1)
-            
+        else:
+            end = dateparser.parse(end).date()
+
         print(f"📡 Buscando edições de {start} até {end}")
         api_url = self._build_api_url(start, end)
         response = requests.get(api_url)
@@ -48,31 +64,37 @@ class ALCEExtractor(ExtractorBase):
 
     def _download_edition(self, edition: dict):
         data_pub = edition["data_publicacao"][:10]
-        file_name = f"diario-alce-{data_pub}.pdf"
-        caminho = edition["caminho_documento_pdf"]
-        url_pdf = self.base_url + caminho
+        filename = f"diario-alce-{data_pub}.pdf"
+        filepath = self.downloads_dir / filename
 
-        print(f"📄 Baixando edição de {data_pub}: {file_name}")
+        if filepath.exists():
+            print(f"⏭️ [{data_pub}] Já existe, pulando.")
+            return True
 
-        response = requests.get(url_pdf, timeout=15)
+        print(f"📄 Baixando edição de {data_pub}: {filename}")
+
+        response = requests.get(filepath, timeout=15)
         if response.status_code == 200 and b"%PDF" in response.content[:10]:
-            local_path = self.downloads_dir / file_name
+            local_path = self.downloads_dir / filename
             with open(local_path, "wb") as f:
                 f.write(response.content)
 
             file_hash = hashlib.md5(response.content).hexdigest()
-            self._save_metadata(file_name, url_pdf, local_path, 'pdf' ,file_hash)
-            print(f"✅ Salvo: {file_name} | Hash: {file_hash[:8]}")
+            self._save_metadata(filename, filepath, local_path, "pdf", file_hash)
+            print(f"✅ Salvo: {filename} | Hash: {file_hash[:8]}")
         else:
-            print(f"⚠️ Arquivo inválido ou não encontrado: {url_pdf}")
+            print(f"⚠️ Arquivo inválido ou não encontrado: {filepath}")
+
 
 if __name__ == "__main__":
     extractor = ALCEExtractor()
-    
+
     start_date = datetime.date(2025, 5, 22)
     end_date = datetime.now().date()
 
-    print(f"🚀 Iniciando download de diários oficiais da AL-CE de {start_date} a {end_date}")
+    print(
+        f"🚀 Iniciando download de diários oficiais da AL-CE de {start_date} a {end_date}"
+    )
 
     extractor.download(start_date, end_date)
